pyproject.toml
[tool.poetry]
name = "276 flash-card-project-start"
version = "0.1.0"
description = ""
authors = ["ivanhow <71693631+ivanhow@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.0"
requests = "^2.27.1"
pysftp = "^0.2.9"
Pillow = "^9.0.1"
pyperclip = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
